Visualization of alert data using chart.js with routing for each alert
